backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
requests==2.31.0
python-dotenv==1.0.0
shodan==1.29.1